Astreriobid analytics: import common code between two adapters (#13320)
* Expand AGENTS guidelines

* Update AGENTS.md

* Update AGENTS.md

* Update AGENTS.md

* Update AGENTS.md

* Update AGENTS.md

* Update AGENTS.md

* Update AGENTS.md

* Update AGENTS.md

* Update AGENTS.md

* Update AGENTS.md

* Update AGENTS.md

* Update AGENTS.md

* Update AGENTS.md

* Update AGENTS.md

* Create CLAUDE.md

* Update AGENTS.md

* adapter: import common asteriobid helpers
